renamed config.py to config_handler.py and created separate project_config.py and plot_config.py modules
